requirements.txt
#tensorflow==1.15.0
tensorflow-cpu
#streamlit==0.67.0
streamlit
numpy==1.19.2
tensorflow_hub
#tensorflow.keras.models


